Several changes: - Fixed comments in filtration_function_controller - Fixed typos in validation drilling validation messages - Added drilling fluids composition validation to filtration function - Fixed MudComposition so it saves and retrieves strings in the Component column - Added frontend validation and error modal frontend to filtration function edit page - Removed unneded console logs in modal error frontend file - Added more modal error frontend title cases (section along with tab)

+++ app/MudComposicion.php
@@ -6,38 +6,35 @@ use Illuminate\Database\Eloquent\Model;
 
 class MudComposicion extends Model
 {
-    protected $table = 'mud_composicion';
-    public $timestamps = false;
+	protected $table = 'mud_composicion';
+	public $timestamps = false;
 
-     protected $fillable = ['d_filtration_function_id','component','concentraction'];
+	protected $fillable = ['d_filtration_function_id', 'component', 'concentraction'];
 
-     static function store($request, $filtration_function){
-     	if($filtration_function->composicion){
-            MudComposicion::where('d_filtration_function_id', $filtration_function->id)->delete();
-          }
+	static function store($request, $filtration_function)
+	{
+		if ($filtration_function->composicion) {
+			MudComposicion::where('d_filtration_function_id', $filtration_function->id)->delete();
+		}
 
-     	$data = json_decode(str_replace(",[null,null]","",$request));
-		foreach($data as $datos)
-		{
+		$data = json_decode(str_replace(",[null,null]", "", $request));
+		foreach ($data as $datos) {
 			$store = new MudComposicion;
 			$store->d_filtration_function_id = $filtration_function->id;
-			$store->component = (float)str_replace(",", ".", $datos[0]);
-			$store->concentraction = (float)str_replace(",", ".", $datos[1]);
+			$store->component = str_replace(",", ".", $datos[0]);
+			$store->concentraction = (float) str_replace(",", ".", $datos[1]);
 			$store->save();
-
 		}
-     }
+	}
 
-    static function getData($id)
-    {
-    	
-    	$datos = collect([]);
+	static function getData($id)
+	{
+		$datos = collect([]);
 
-    	foreach(MudComposicion::where('d_filtration_function_id', $id)->get() as $mudComposicion) 
-    	{
-    		$datos->push([(float)$mudComposicion->component, (float)$mudComposicion->concentraction]);
-    	}
+		foreach (MudComposicion::where('d_filtration_function_id', $id)->get() as $mudComposicion) {
+			$datos->push([$mudComposicion->component, (float) $mudComposicion->concentraction]);
+		}
 
-    	return $datos;
-    }
+		return $datos;
+	}
 }
